refactor(gui): move PrimerDimerView to primer_dimer package
Isolate PrimerDimerView and extract the primer dimer context card
UI component into its own dedicated class/file for enhanced code reuse.
