+++ src/amplifyp/gui/views/primer_dimer/primer_dimer_view.py
@@ -1,0 +1,102 @@
+# Copyright (C) 2026 Fufu Fang
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+"""Primer Dimer View for the Flet application."""
+
+import traceback
+
+import flet as ft
+
+from amplifyp.dimer import PrimerDimerGenerator
+from amplifyp.dna import Primer
+from amplifyp.gui.settings import GUIColors, GUISettings
+from amplifyp.gui.user_data import GUIInput
+from amplifyp.gui.util import show_error_dialog
+from amplifyp.gui.views.primer_dimer.primer_dimer_card import PrimerDimerCard
+
+
+class PrimerDimerView(ft.Column):  # type: ignore[misc]
+    """View to show calculated primer dimer results."""
+
+    def __init__(
+        self,
+        page: ft.Page,
+        input_data: GUIInput | None = None,
+        settings: GUISettings | None = None,
+    ) -> None:
+        """Initialize the PrimerDimerView."""
+        super().__init__(expand=True)
+        self.app_page = page
+        self.input_data = input_data if input_data is not None else GUIInput()
+        self.settings = settings if settings is not None else GUISettings()
+
+        self.result_list = ft.ListView(
+            expand=True, spacing=10, scroll=ft.ScrollMode.ALWAYS
+        )
+        self.controls = [
+            ft.Container(content=self.result_list, expand=True),
+        ]
+
+    def run_analysis(self) -> bool:
+        """Run primer dimer analysis and populate the UI."""
+        self.result_list.controls.clear()
+        success = True
+        try:
+            pd_settings = self.settings.get_primer_dimer_settings()
+            generator = PrimerDimerGenerator(settings=pd_settings)
+
+            primers = self.input_data.get_active_primers()
+            for p in primers:
+                name = p["name"]
+                seq = p["seq"]
+                generator.add_primer(Primer(sequence=seq, name=name))
+
+            generator.analyse_primers()
+
+            dimers = generator.primer_dimers
+
+            if not dimers:
+                self.result_list.controls.append(
+                    ft.Container(
+                        content=ft.Text(
+                            "No primer dimers detected above threshold.",
+                            size=self.settings.get("font_size_subheader", 16),
+                            italic=True,
+                            text_align=ft.TextAlign.CENTER,
+                        ),
+                        alignment=ft.Alignment(0, 0),
+                        padding=20,
+                    )
+                )
+            else:
+                font_family = self.settings.get("font_family", "Roboto Mono")
+                for d in dimers:
+                    card = PrimerDimerCard(
+                        d,
+                        self.settings,
+                        font_family=font_family,
+                    )
+                    self.result_list.controls.append(card)
+        except Exception as ex:
+            self.result_list.controls.append(
+                ft.Text(
+                    f"Error running analysis: {ex}\n{traceback.format_exc()}",
+                    color=GUIColors.ERROR_RED,
+                )
+            )
+            show_error_dialog(self.app_page, "Error running analysis", str(ex))
+            success = False
+        self.app_page.update()
+        return success
